updated some files

+++ src/dvd.cpp
@@ -4,7 +4,7 @@ DVD::DVD(DVDType type, int stock, const std::string& title, const std::string& d
          int releaseYear, int rentalPeriodInDays): Rentable(RentableType::DVD,
 		 stock, title, releaseYear, rentalPeriodInDays)
 {
-	this->subtype = type;
+	this->subtype = (int)type;
 	this->director = director;
 }
